Writes the correct number of records written to the database to CloudWatch.

+++ src/main/java/uk/gov/indexer/monitoring/CloudwatchRecordsProcessedUpdater.java
@@ -25,7 +25,7 @@ public class CloudwatchRecordsProcessedUpdater {
         acw.setEndpoint("https://monitoring.eu-west-1.amazonaws.com");
     }
 
-    public void update(int recordsProcessed) {
+    public void update(long recordsProcessed) {
         acw.putMetricData(new PutMetricDataRequest()
                 .withNamespace(environment)
                 .withMetricData(Collections.singletonList(new MetricDatum()
